# ui/openapi/fitnessclub.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FitnessClub.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/trainers": {
      "get": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_List",
        "parameters": [
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainerSummaryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainerSummaryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainerSummaryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_Hire",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainerRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainerRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrainerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/trainers/{id}": {
      "get": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_UpdateProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainerRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainerRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrainerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/trainers/{id}/activate": {
      "post": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_Activate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/trainers/{id}/deactivate": {
      "post": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_Deactivate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/trainers/{id}/working-hours": {
      "put": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_SetWorkingHours",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WorkingHoursRequest"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WorkingHoursRequest"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WorkingHoursRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/trainers/{id}/clients/{clientId}": {
      "post": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_AssignClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_UnassignClient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/trainers/{id}/identity": {
      "put": {
        "tags": [
          "Trainers"
        ],
        "operationId": "Trainers_LinkIdentity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkIdentityRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkIdentityRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LinkIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DayOfWeek": {
        "enum": [
          "Sunday",
          "Monday",
          "Tuesday",
          "Wednesday",
          "Thursday",
          "Friday",
          "Saturday"
        ]
      },
      "LinkIdentityRequest": {
        "required": [
          "identityUserId"
        ],
        "type": "object",
        "properties": {
          "identityUserId": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "TrainerClientResponse": {
        "required": [
          "clientId",
          "fullName",
          "assignedAt"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TrainerRequest": {
        "required": [
          "firstName",
          "lastName",
          "phone",
          "specialization"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "lastName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "middleName": {
            "maxLength": 100,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "maxLength": 32,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "maxLength": 254,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "specialization": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "TrainerResponse": {
        "required": [
          "id",
          "firstName",
          "lastName",
          "middleName",
          "fullName",
          "phone",
          "email",
          "specialization",
          "isActive",
          "identityUserId",
          "workingHours",
          "clients"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": [
              "null",
              "string"
            ]
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "specialization": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "identityUserId": {
            "type": [
              "null",
              "string"
            ]
          },
          "workingHours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingHoursResponse"
            }
          },
          "clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrainerClientResponse"
            }
          }
        }
      },
      "TrainerSummaryResponse": {
        "required": [
          "id",
          "firstName",
          "lastName",
          "middleName",
          "fullName",
          "phone",
          "email",
          "specialization",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": [
              "null",
              "string"
            ]
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "specialization": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "WorkingHoursRequest": {
        "required": [
          "day",
          "start",
          "end"
        ],
        "type": "object",
        "properties": {
          "day": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DayOfWeek"
              }
            ]
          },
          "start": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "end": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          }
        }
      },
      "WorkingHoursResponse": {
        "required": [
          "day",
          "start",
          "end"
        ],
        "type": "object",
        "properties": {
          "day": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "start": {
            "type": "string",
            "format": "time"
          },
          "end": {
            "type": "string",
            "format": "time"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Auth0 access token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}